+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.0.0] - 2020-08-19
 ### Fixed
 - Make sure QrCode `min_amount` and `max_amount` make sense
 - Fix call to undefined method negative when using refundByMerchantItemId